requirements.txt
# requirements.txt
wmi==1.5.1
GPUtil==1.4.0
psutil==5.9.4
pywin32==305
pystray==0.19.3
Pillow==9.5.0
win10toast==0.9
configparser==5.3.0